requirement.txt
alembic==1.4.0
autopep8==1.5
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
httpie==2.0.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.1
MarkupSafe==1.1.1
pkg-resources==0.0.0
psycopg2-binary==2.8.4
pycodestyle==2.5.0
Pygments==2.5.2
python-dateutil==2.8.1
python-dotenv==0.11.0
python-editor==1.0.4
requests==2.22.0
selenium==3.141.0
six==1.14.0
soupsieve==1.9.5
SQLAlchemy==1.3.13
urllib3==1.25.8
Werkzeug==1.0.0
WTForms==2.2.1
